ISSUE-128: refactored to prevent duplicate code
