+++ gdax/authenticated_client.py
@@ -12,6 +12,7 @@ import base64
 import json
 from requests.auth import AuthBase
 from gdax.public_client import PublicClient
+from gdax.gdax_auth import GdaxAuth
 
 
 class AuthenticatedClient(PublicClient):
@@ -289,27 +290,3 @@ class AuthenticatedClient(PublicClient):
         r = requests.get(self.url + "/users/self/trailing-volume", auth=self.auth, timeout=30)
         # r.raise_for_status()
         return r.json()
-
-
-class GdaxAuth(AuthBase):
-    # Provided by gdax: https://docs.gdax.com/#signing-a-message
-    def __init__(self, api_key, secret_key, passphrase):
-        self.api_key = api_key
-        self.secret_key = secret_key
-        self.passphrase = passphrase
-
-    def __call__(self, request):
-        timestamp = str(time.time())
-        message = timestamp + request.method + request.path_url + (request.body or '')
-        message = message.encode('ascii')
-        hmac_key = base64.b64decode(self.secret_key)
-        signature = hmac.new(hmac_key, message, hashlib.sha256)
-        signature_b64 = base64.b64encode(signature.digest())
-        request.headers.update({
-            'Content-Type': 'Application/JSON',
-            'CB-ACCESS-SIGN': signature_b64,
-            'CB-ACCESS-TIMESTAMP': timestamp,
-            'CB-ACCESS-KEY': self.api_key,
-            'CB-ACCESS-PASSPHRASE': self.passphrase
-        })
-        return request
