+++ gdax/gdax_auth.py
@@ -1,0 +1,34 @@
+import hmac
+import hashlib
+import time
+import base64
+from requests.auth import AuthBase
+
+
+class GdaxAuth(AuthBase):
+    # Provided by gdax: https://docs.gdax.com/#signing-a-message
+    def __init__(self, api_key, secret_key, passphrase):
+        self.api_key = api_key
+        self.secret_key = secret_key
+        self.passphrase = passphrase
+
+    def __call__(self, request):
+        timestamp = str(time.time())
+        message = timestamp + request.method + request.path_url + (request.body or '')
+        request.headers.update(get_auth_headers(timestamp, message, self.api_key, self.secret_key,
+                                                self.passphrase))
+        return request
+
+
+def get_auth_headers(timestamp, message, api_key, secret_key, passphrase):
+    message = message.encode('ascii')
+    hmac_key = base64.b64decode(secret_key)
+    signature = hmac.new(hmac_key, message, hashlib.sha256)
+    signature_b64 = base64.b64encode(signature.digest()).decode('utf-8')
+    return {
+        'Content-Type': 'Application/JSON',
+        'CB-ACCESS-SIGN': signature_b64,
+        'CB-ACCESS-TIMESTAMP': timestamp,
+        'CB-ACCESS-KEY': api_key,
+        'CB-ACCESS-PASSPHRASE': passphrase
+    }
